Backend (python)/OMR Evaluation API/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.0
opencv-python==4.9.0.80
numpy==1.26.4
FPDF2==2.7.8
Pillow==10.2.0
Gunicorn
